+++ src/app/access-control/access-control.controller.ts
@@ -1,14 +1,14 @@
 import { Body, Controller, Delete, Get, Param, Post, Put, Req, UseGuards } from '@nestjs/common';
+import { ApiTags } from '@nestjs/swagger';
 import { AccessControlService } from './access-control.service';
 import { AccessControlInputDto } from './dto/access-control-input.dto';
-import { ApiTags } from '@nestjs/swagger';
-import { CustomRequest } from '../common/interfaces/token.interface';
+import { AccessJwtGuard } from '../../guards/access-jwt.guard';
+import { CustomRequest } from '../../common/interfaces/token.interface';
 import { AccessControlOutputDto } from './dto/access-control-output.dto';
-import { CreateAccessControlSwDecorator } from './decorators/create-access-control-sw.decorator';
-import { UpdateAccessControlSwDecorator } from './decorators/update-access-control-sw.decorator';
-import { DeleteAccessControlSwDecorator } from './decorators/delete-access-control-sw.decorator';
-import { GetAllAccessControlSwDecorator } from './decorators/get-all-access-control-sw.decorator';
-import { AccessJwtGuard } from '../guards/access-jwt.guard';
+import { CreateAccessControlSwDec } from './decorators/create-access-control-sw.decorator';
+import { UpdateAccessControlSwDec } from './decorators/update-access-control-sw.decorator';
+import { DeleteAccessControlSwDec } from './decorators/delete-access-control-sw.decorator';
+import { GetAllAccessControlSwDec } from './decorators/get-all-access-control-sw.decorator';
 
 @UseGuards(AccessJwtGuard)
 @ApiTags('Access control')
@@ -16,21 +16,21 @@ import { AccessJwtGuard } from '../guards/access-jwt.guard';
 export class AccessControlController {
   constructor(private readonly accessControlService: AccessControlService) {}
 
-  @CreateAccessControlSwDecorator()
+  @CreateAccessControlSwDec()
   @Post()
   async create(@Body() dto: AccessControlInputDto, @Req() req: CustomRequest): Promise<AccessControlOutputDto> {
     const { userId } = req.user;
     return this.accessControlService.create(userId, dto);
   }
 
-  @DeleteAccessControlSwDecorator()
+  @DeleteAccessControlSwDec()
   @Delete(':id')
   async delete(@Param('id') accessId: string, @Req() req: CustomRequest): Promise<AccessControlOutputDto> {
     const { userId } = req.user;
     return this.accessControlService.delete(userId, accessId);
   }
 
-  @UpdateAccessControlSwDecorator()
+  @UpdateAccessControlSwDec()
   @Put(':id')
   async update(
     @Param('id') accessId: string,
@@ -41,7 +41,7 @@ export class AccessControlController {
     return this.accessControlService.update(userId, accessId, dto);
   }
 
-  @GetAllAccessControlSwDecorator()
+  @GetAllAccessControlSwDec()
   @Get()
   async get(@Req() req: CustomRequest): Promise<AccessControlOutputDto[]> {
     const { userId } = req.user;
