+++ src/app/access-control/decorators/is-optional-mongoid-array.decorator.ts
@@ -1,6 +1,6 @@
 import { applyDecorators } from '@nestjs/common';
 import { ArrayNotEmpty, IsArray, IsMongoId, IsOptional } from 'class-validator';
 
-export function IsOptionalMongoidArray() {
+export function IsOptionalMongoidArrayDec() {
   return applyDecorators(IsOptional(), IsArray(), ArrayNotEmpty(), IsMongoId({ each: true }));
 }
